sha256

+++ crypto/sha256.c
@@ -1,5 +1,5 @@
 #include "hblk_crypto.h"
-uint8_t *sha256(int8_t const *s, size_t  len,
+uint8_t *sha256(int8_t const *s, unsigned int len,
 uint8_t digest[SHA256_DIGEST_LENGTH])
 {
 if (s)
